+++ paddlenlp/utils/__init__.py
@@ -23,6 +23,7 @@ from .infohub import infohub
 from .initializer import to
 from .memory_utils import empty_device_cache
 from .optimizer import *
+from .paddle_patch import *
 from .serialization import load_torch
 
 # hack impl for EagerParamBase to function
